added serialization

+++ src/activation.rs
@@ -1,19 +1,21 @@
-use std::f64::consts::E;
+use std::f32::consts::E;
 
-#[derive(Clone, Copy)]
+use serde::{Serialize, Deserialize};
+
+#[derive(Serialize, Deserialize, Clone, Copy)]
 pub enum Act {
     Tanh,
     Sig,
     Lin
 }
 
-fn sigmoid(n: f64) -> f64 {
+fn sigmoid(n: f32) -> f32 {
     1.0 / (1.0 + E.powf(-n))
 }
 
 impl Act {
     /// Applies non-linearity function to `n`
-    pub fn value(&self, n: f64) -> f64 {
+    pub fn value(&self, n: f32) -> f32 {
         match self {
             Act::Tanh => n.tanh(),
             Act::Sig  => sigmoid(n),
@@ -22,7 +24,7 @@ impl Act {
     }
 
     /// Applies non-linearity derivative to `n`
-    pub fn deriv(&self, n: f64) -> f64 {
+    pub fn deriv(&self, n: f32) -> f32 {
         match self {
             Act::Tanh => 1.0  - n.tanh().powi(2),
             Act::Sig  => {
