+++ src/cost.rs
@@ -1,19 +1,21 @@
+use serde::{Serialize, Deserialize};
 
-#[derive(Clone, Copy)]
+
+#[derive(Serialize, Deserialize, Clone, Copy)]
 pub enum Cost {
     MSE
 }
 
 impl Cost {
     /// Applies cost function to `err`
-    pub fn value(&self, err: f64) -> f64 {
+    pub fn value(&self, err: f32) -> f32 {
         match self {
             Cost::MSE => err.powi(2)
         }
     }
 
     /// Applies cost derivative to `err`
-    pub fn deriv(&self, err: f64) -> f64 {
+    pub fn deriv(&self, err: f32) -> f32 {
         match self {
             Cost::MSE => 2.0 * err
         }
